restructure code in pkg/dryrun
